Pennsylvania Breeding Survey Data (BBS) from the USGS.

A dataset containing Breeding Bird Survey (BBS) data collected by the USGS since 1967. Some columns have been removed to simplify the data. See ftp://ftpext.usgs.gov/pub/er/md/laurel/BBS/DataFiles/Summary.txt for the original metadata file.

BBS_PA

Format

A data frame with 53940 rows and 10 variables:

countrynum

Numeric code indicating country. USA = 840.

statenum

Numeric code indicating state. PA = 72.

Route

numeric code indicating router number

RPID

Run Protocol Identification Number. See RunProtocolID.txt file in USGS documentation

Year

Yeardata collected.

Aou

numeric species code from American Ornithological Union

StopTotal

Total number of stops out of 50 on which the species was recorded. This is not a count of birds, but the number of stops that had a particular bird! The min value is 1, and the max value is 50

SpeciesTotal

Total number of stops out of 50 on which the species was recorded. This is what is usually used for modeling. The min value is 1 because a given species in a given year only occurs in the data if it was observed; there are therefore no zeros. These can be added using the expandBBS_routes() function.

Source

www.pwrc.usgs.gov/BBS/RawData

References

Pardieck, K.L., D.J. Ziolkowski Jr., M. Lutmerding, K. Campbell and M.-A.R. Hudson. 2017. North American Breeding Bird Survey Dataset 1966 - 2016, version 2016.0. U.S. Geological Survey, Patuxent Wildlife Research Center. doi:10.5066/F7W0944J.

Rosenberg et al. 2017. Use of North American Breeding Bird Survey data in avian conservation assessments The Condor. http://www.bioone.org/doi/abs/10.1650/CONDOR-17-57.1