Brook trout impacts on Chinook salmon survival - original data

Levin et al. 2002 studied the impacts of brook trout introduced for sport fishing on native Chinook salmon. Whitlock & Shulter used a subset of these data as an example in their text Analysis of Biological Data

brook_trout

Format

A data frame, including:

site

Name of study site or stream

trout.PRES.ABS

Taxonomic family

release

Number of Chinook salmon released

recap

Number of juvenil chinook salmon "recapatured" by detecting their PIT tag; this was not reported in original paper but was estimated using published survival rate. The survival rate estiamted using mark-recapture methods to account for detection probabilities.

surv

survival rate

Source

http://rspb.royalsocietypublishing.org/content/269/1501/1663.short

References

Levin et al. 2002. Non–indigenous brook trout and the demise of Pacific salmon: a forgotten threat? PRSB 269. DOI: 10.1098/rspb.2002.2063

Whitlock & Schulter. 2nd ed. Analysis of Biological Data.

Examples

t.test(surv ~ trout.PRES.ABS, data = brook_trout)

#> 
#> 	Welch Two Sample t-test
#> 
#> data:  surv by trout.PRES.ABS
#> t = 2.5974, df = 53.693, p-value = 0.0121
#> alternative hypothesis: true difference in means is not equal to 0
#> 95 percent confidence interval:
#>  0.0133692 0.1038965
#> sample estimates:
#>  mean in group abs mean in group pres 
#>          0.2208084          0.1621756 
#>