Mammal milk dataset

A dataset on the nutritional composition of mammalian milk.

milk

Format

A data frame with 23 columns, including:

order

Taxonomic order

family

Taxonomic family

spp

Species

mass.female

Mean mass of female

gestation.month

Duration of gestation in months (how long pregnant)

lacatation.months

duration of lacation in months (how long until offspring weaned)

mass.litter

Total mass of litter (all offspring combined)

repro.output

Number of offspring

dev.stage.at.birth

Developemental stage at birth

diet

Diet of organism

arid

Does it live in a desert or other arid biome?

biome

Type of biome

N

Number of...?

lactation.stage.orig

?

dry.matter

Dry matter content of milk

fat

Fat content of milk

fat.revised

?

protein

Protein content of milk

sugar

suger content of milk

energy

energetic value of milk

ref

refernece for original source of data on this species

gest.month.NUM

...

lacat.mo.NUM

...

Source

http://www.diamondse.info/

References

Skibiel et al. 2013. The evolution of the nutrient composition of mammalian milks. Journal of Animal Ecology. 82:1254–1264 DOI: 10.1111/1365-2656.12095 http://onlinelibrary.wiley.com/doi/10.1111/1365-2656.12095/full